requirements.txt
anthropic
asyncssh
pyyaml
rich
prompt_toolkit
ollama
google-generativeai
openai
keyring
pydantic
